// frontend/src/api/generated/models/DownloadJobItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DownloadJobItem = {
    jobId: number;
    bookMd5: string;
    format: string;
    status: string;
    progress: number;
    filePath?: string | null;
    error?: string | null;
    createdAt: string;
    updatedAt: string;
};